# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/referentiels": {
      "get": {
        "tags": ["Referentiels"],
        "summary": "Get all referentiels",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "List of referentiels"}
        }
      },
      "post": {
        "tags": ["Referentiels"],
        "summary": "Create a referentiel",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/Referentiel"}
            }
          }
        },
        "responses": {
          "201": {"description": "Referentiel created successfully"}
        }
      }
    },
...
    "/api/v1/apprenants": {
      "post": {
        "tags": ["Apprenants"],
        "summary": "Create a new apprenant",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/Apprenant"}
            }
          }
        },
        "responses": {
          "201": {"description": "Apprenant created successfully"}
        }
      },
      "get": {
        "tags": ["Apprenants"],
        "summary": "Get all apprenants",
        "security": [{"bearerAuth": []}],
        "responses": {
          "200": {"description": "List of apprenants"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Referentiel": {
        "type": "object",
        "properties": {
          "name": {"type": "string"}
        }
      },
...
      "Apprenant": {
        "type": "object",
        "properties": {
          "first_name": {"type": "string"},
          "last_name": {"type": "string"},
          "email": {"type": "string"}
        }
      }
    }
  }
}